+++ PSModule/ModernWorkplaceClientCenter/Functions/Get-MDMEnrollmentStatus.ps1
@@ -13,6 +13,8 @@ function Get-MDMEnrollmentStatus {
     param()
     #Locate correct Enrollment Key
     $EnrollmentKey = Get-Item -Path HKLM:\SOFTWARE\Microsoft\Enrollments\* | Get-ItemProperty | Where-Object -FilterScript {$null -ne $_.UPN}
-    Add-Member -InputObject $EnrollmentKey -MemberType NoteProperty -Name EnrollmentTypeText -Value (Invoke-TranslateMDMEnrollmentType -Id ($EnrollmentKey.EnrollmentType))
+    if($EnrollmentKey){
+        Add-Member -InputObject $EnrollmentKey -MemberType NoteProperty -Name EnrollmentTypeText -Value (Invoke-TranslateMDMEnrollmentType -Id ($EnrollmentKey.EnrollmentType))
+    }
     return $EnrollmentKey
 }